% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate.DET}
\alias{annotate.DET}
\title{This function adds a geom to a plot which annotates DET curves with their names.
The annotations are spaced apart so as not to occlude one another.}
\usage{
annotate.DET(stat = "identity", data = NULL, mapping = aes(FMR, FNMR,
  label = label), position = "identity", ..., inherit.aes = TRUE,
  show.legend = FALSE, bgfill = "white", bgalpha = 1,
  label.padding = unit(0.25, "lines"), na.rm = FALSE, parse = FALSE)
}
\description{
See the annotation function in ggplot2 for the meaning of the parameters.
}
\examples{
gen.scores <- runif(1e5, 0, 10)
imp.scores <- rnorm(1e5)
DET <- make.DET(gen.scores, imp.scores)
DET[, curve := "Curve Name"]
plot.DET(DET, xlim=c(1e-4, 0.2), col="darkred", legend.position='none') +
annotate.DET()
}
