% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors.15}
\alias{colors.15}
\title{DET plotting colors.}
\format{A vector of character strings}
\usage{
colors.15
}
\description{
Chosen and ordered such that similar colours are spaced apart from one another.
}
\keyword{datasets}
