% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{DET.theme}
\alias{DET.theme}
\title{Sets the default theme for DET plots.}
\usage{
DET.theme(panel.grid.minor = element_line(colour = "grey90", size = 0.2),
  panel.grid.major = element_line(colour = "grey80", size = 0.3),
  axis.text = element_text(colour = "darkblue", size = 10),
  axis.title = element_text(colour = "darkblue"),
  axis.line = element_line(colour = "darkblue"),
  legend.text = element_text(colour = "darkblue"),
  legend.key.height = unit(0.25, "cm"), legend.position = "right",
  legend.title = element_blank(), plot.title = element_blank(),
  plot.margin = unit(c(0.2, 0.5, 0.1, 0.1), "cm"),
  panel.border = element_rect(colour = "darkblue"), ...)
}
\description{
Sets the default theme for DET plots.
}
