% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{GeomDetAnnotate}
\alias{GeomDetAnnotate}
\title{The Geom that draws the curve annotations, called by annotate.DET().}
\format{An object of class \code{GeomDetAnnotate} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomDetAnnotate
}
\description{
The Geom that draws the curve annotations, called by annotate.DET().
}
\keyword{datasets}
