% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det.R
\name{get.FNMR}
\alias{get.FNMR}
\title{Get FNMR at the specified FMR(s) or threshold(s).}
\usage{
get.FNMR(DET, ...)
}
\arguments{
\item{DET}{A data.table object or path to a DET file.}

\item{FMR}{A single false match rate (or vector of FMRs).}

\item{t}{A single decision threshold (or vector of thresholds).}
}
\value{
A vector of false non-match rates.
}
\description{
Must specify either FMR or t but not both.
}
