% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det.R
\name{get.threshold}
\alias{get.threshold}
\title{Get threshold at the specified FMR or FNMR.}
\usage{
get.threshold(DET, ...)
}
\arguments{
\item{DET}{A data.table object or path to a DET file.}

\item{FMR}{A single false match rate (or vector of FMRs).}

\item{FNMR}{A singe false match rate (or vector of FNMRs).}
}
\value{
A vector of thresholds.
}
\description{
Get threshold at the specified FMR or FNMR.
}
