% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{log.scale}
\alias{log.scale}
\title{Returns a scale object for continuous x or y aesthetics.}
\usage{
\method{log}{scale}(..., expand = c(0.02, 0), trans = "log10",
  breaks = log.ticks(), minor_breaks = log.ticks(loc = 1:9),
  labels = scientific.notation, axis = "x")
}
\arguments{
\item{expand}{See scale_x_continuous() in ggplot2 package.}

\item{trans}{See scale_x_continuous() in ggplot2 package.}

\item{breaks}{See scale_x_continuous() in ggplot2 package.}

\item{minor_breaks}{See scale_x_continuous() in ggplot2 package.}

\item{labels}{See scale_x_continuous() in ggplot2 package.}

\item{axis}{"x" for horizontal axis (default), "y" for vertical axis.}

\item{breaks}{See scale_x_continuous() in ggplot2 package.}
}
\description{
Returns a scale object for continuous x or y aesthetics.
}
