% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{log.ticks}
\alias{log.ticks}
\title{Returns conveniently located tick marks for a log10 plot.}
\usage{
\method{log}{ticks}(lim = c(-10, 1), loc = 1)
}
\arguments{
\item{lim}{log10 of the axis limits.}

\item{loc}{The locations of the tick marks for each factor of 10.}
}
\value{
The locations of the tick marks.
}
\description{
Returns conveniently located tick marks for a log10 plot.
}
