% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det.R
\name{make.DET}
\alias{make.DET}
\title{Constructs a DET object from a set of mated and nonmated scores.}
\usage{
make.DET(mated.scores, nonmated.scores, FMR.range = c(1e-08, 1),
  num.points = 4096, similarity = "auto", positive.system = TRUE,
  one.to.one = TRUE)
}
\arguments{
\item{nonmated.scores}{impostor (aka nonmated) comparison scores.}

\item{FMR.range}{range over which to construct the DET.}

\item{num.points}{The number of decision thresholds to use when generating the DET
curve.}

\item{similarity}{Are the scores measures of similarity (TRUE) or distances (FALSE).}

\item{gen.scores}{genuine (aka mated) comparison scores.}

\item{positive}{Is it a positive or negative system (determines how FTEs are treated).}
}
\value{
A DET curve as a data.table with three columns: "t", "FMR", and "FNMR".
}
\description{
If num.points is a single numerical value, then it specifies the number of decision
thresholds to use when generating the DET curve. The selected thresholds will
correspond to evenly spaced FMRs in the log space. If num.points is a numeric vector,
then the elements of the vector will be used as the decision thresholds when
generating the DET curve.
}
\details{
Failure-to-enroll (FTE) comparisons should be represented by NAs. When positive is
set to true, FTE comparisons contribute to the false non-match rate. When
positive is set to false, FTE comparisons contribute to the false match rate.
}
\examples{
gen.scores <- rnorm(1e5, mean=4)
nonmated.scores <- rnorm(1e5, mean=0)
DET <- make.DET(gen.scores, nonmated.scores)
get.FNMR(DET, FMR=1e-2)
}
