% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.DET}
\alias{plot.DET}
\title{Plots one or more DET curves.}
\usage{
\method{plot}{DET}(DETs, xlim = c(1e-06, 0.2), ylim = "auto",
  xlabel = intersect(c("FMR", "FPIR"), names(DETs)),
  ylabel = intersect(c("FNMR", "FNIR"), names(DETs)),
  x.axis = log.scale(name = xlabel, axis = "x"),
  y.axis = log.scale(name = ylabel, axis = "y", labels = identity, breaks
  = log.ticks(loc = c(1, 2, 5))), prepend.FNMR = FALSE,
  FMR.sort = mean(xlim), plotter = geom_step,
  legend = scale_colour_manual(values = colors.15), colour = key(DETs),
  theme = DET.theme, legend.position = "auto", ...)
}
\arguments{
\item{DETs}{A data.table representing the DET curve(s) to plot.}

\item{xlim}{The horizontal axis limits (default: c(8e-7, 0.2)).}

\item{ylim}{The vertical axis limits. "auto" (the default) selects limits that
fully display all curves over the range specified by xlim.}

\item{xlabel}{The horizontal axis label (default: FMR)}

\item{ylabel}{The vertical axis label (default: FNMR)}

\item{x.axis}{The horizontal axis scale object.}

\item{y.axis}{The vertical axis scale object.}

\item{FMR.sort}{Sort the curves in the legend according to their FNMR at this
FMR (default: 1e-5).}

\item{plotter}{The ggplot2 line drawing function (default: geom_step).}

\item{theme}{The ggplot2 theme to apply (default: DET.theme()).}

\item{...}{Additional parameters to pass to the theme object.}

\item{print.FNMR}{If nonzero, prepend legend labels with the curve's FNMR at
FMR=print.FNMR. If zero (the default), nothing is prepended.}
}
\value{
A ggplot plotting object.
}
\description{
Plots one or more DET curves.
}
\examples{
gen.scores <- runif(1e5, 0, 10)
imp.scores <- rnorm(1e5)
DET <- make.DET(gen.scores, imp.scores)
plot.DET(DET, xlim=c(1e-4, 0.2), col="darkred", legend.position='none')
}
