% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scientific.notation}
\alias{scientific.notation}
\title{Prints a number in human-interpretable exponential format.}
\usage{
scientific.notation(number)
}
\arguments{
\item{number}{The numerical value to print.}
}
\value{
The number in scientific notation as an expression object.
}
\description{
Prints a number in human-interpretable exponential format.
}
